refactor: move parseConnectionString method to utils
